refactor data table stories directory structure
